# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  controllers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.RequestPasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  controllers.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - code
    - email
    - new_password
    type: object
...
  github_com_leleo886_lopic_internal_error.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.Role:
    properties:
      allowed_extensions:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      gallery_open:
        type: boolean
      id:
        type: integer
      max_albums_per_user:
        type: integer
      max_file_size_mb:
        type: integer
      max_files_per_upload:
        type: integer
      max_storage_size_mb:
        type: integer
      name:
        type: string
      storage_name:
        description: 存储配置名称
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      image_count:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      total_size:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  services.LoginResponse:
    properties:
      token_response:
        $ref: '#/definitions/services.TokenResponse'
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  services.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
  success.DataResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  success.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API
  title: Lopic API
  version: 1.0.0
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取JWT令牌和刷新令牌
      parameters:
      - description: 用户登录信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回用户信息和访问令牌和刷新令牌
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: 将刷新令牌添加到黑名单，防止再次使用
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 用户登出
      tags:
      - 认证
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌和刷新令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的访问令牌和刷新令牌
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 刷新访问令牌
      tags:
      - 认证
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户，默认获得user角色
      parameters:
      - description: 用户注册信息（默认user角色）
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 用户注册
      tags:
      - 认证
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: 使用验证码设置新密码
      parameters:
      - description: 邮箱、验证码和新密码
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 重置密码
      tags:
      - 认证
  /api/auth/reset-password/request:
    post:
      consumes:
      - application/json
      description: 发送密码重置邮件
      parameters:
      - description: 邮箱地址
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RequestPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 请求密码重置
      tags:
      - 认证
  /api/auth/verify-email:
    get:
      consumes:
      - application/json
      description: 使用验证令牌激活用户账号
      parameters:
      - description: 验证令牌
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      summary: 邮箱验证
      tags:
      - 认证
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"